add title tests to validator-ajv8 validator tests
